Move prometheus helpers to own package
These are going to be need from multiple places within influx-spout.
